data-training-app/djangoProject/requirements.txt
psycopg==3.1.10
psycopg2==2.9.7
numpy~=1.25.2
numba~=0.58.0
Django~=4.2.4
djangorestframework~=3.14.0
matplotlib~=3.8.0
keras~=2.14.0
keras-core
graypy~=2.1.0
tensorflow
scipy~=1.11.3
fastdtw~=0.3.4
django_prometheus
django-cors-headers
drf_yasg
djangorestframework-simplejwt
timedelta
pyjwt
jwcrypto~=1.5.0
pandas~=2.0.3
channels~=4.0.0
channels-redis
asgiref~=3.7.2
daphne~=4.0.0
scikit-learn~=1.3.0
emd~=0.6.2
jsonschema~=4.19.0
statsmodels~=0.14.0
h5py
python-dateutil~=2.8.2
prometheus-client
psutil